New Line Configuration
Remove the telnet and serial newline options and replace them with a global option.
The new option includes CR, LF, CR/LF, a default value, which maps to a default value specific to protocol:
  Telnet and Raw default to CR/LF, others default to CR.
Continue to support the telnet specific option when loading config, and map it to the new option.
Unify newline handling across all backends, so that ldisc.c doesn't have to know any backend specific details.